+++ beacon-chain/rpc/eth/shared/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@prysm//tools/go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["structs.go"],
+    importpath = "github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/shared",
+    visibility = ["//visibility:public"],
+)
