+++ beacon-chain/rpc/eth/validator/structs.go
@@ -1,0 +1,7 @@
+package validator
+
+import "github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/shared"
+
+type AggregateAttestationResponse struct {
+	Data shared.Attestation `json:"data"`
+}
